меню для пользователя и календарь

+++ app/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, DateTime, Index, ForeignKey, Boolean, Date, BigInteger, Integer  # Добавлен Date
+from sqlalchemy import Column, String, DateTime, Index, ForeignKey, Boolean, Date, BigInteger
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship, DeclarativeBase
@@ -62,12 +62,11 @@ class GroupMember(Base):
     user = relationship("User", back_populates="group_membership")
     group = relationship("Group", back_populates="members")
 
-# Добавьте модель Event в app/db/models.py
 class Event(Base):
     __tablename__ = 'events'
     id = Column(UUID(as_uuid=True), primary_key=True, server_default=func.uuid_generate_v4())
     group_id = Column(UUID(as_uuid=True), ForeignKey('groups.id', ondelete='CASCADE'), nullable=False)
-    created_by_user_id = Column(Integer, ForeignKey('users.telegram_id', ondelete='CASCADE'), nullable=False)  # Только одно определение
+    created_by_user_id = Column(BigInteger, ForeignKey('users.telegram_id', ondelete='CASCADE'), nullable=False)  # Исправлено на BigInteger
     title = Column(String(100), nullable=False)
     description = Column(String(255), nullable=True)
     subject = Column(String(100), nullable=True)
@@ -81,8 +80,8 @@ class Invite(Base):
     __tablename__ = 'groupinvitations'
     id = Column(UUID(as_uuid=True), primary_key=True, server_default=func.uuid_generate_v4())
     group_id = Column(UUID(as_uuid=True), ForeignKey('groups.id', ondelete='CASCADE'), nullable=False)
-    invited_by_user_id = Column(Integer, ForeignKey('users.telegram_id'), nullable=False)
-    invite_token = Column(String(36), unique=True, nullable=False)  # Убедитесь, что длина 36
+    invited_by_user_id = Column(BigInteger, ForeignKey('users.telegram_id'), nullable=False)  # Исправлено на BigInteger
+    invite_token = Column(String(36), unique=True, nullable=False)
     expires_at = Column(Date, nullable=False)
     is_used = Column(Boolean, default=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
